requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
convertdate==2.3.2
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-adminlte3==0.1.6
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
gunicorn==20.1.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pymeeus==0.5.11
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
